+++ CHANGELOG.md
@@ -13,6 +13,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - In `grond export`, it is now possible to suppress exporting events with
   combined location uncertainty below a given threshold.
 - Support offset coordinates in station and event inputs.
+- lockfile `.running` in rundir during inverion. Fixes timeout for monitor.
 
 ### Fixed
 - Fix search in HTML report.